add set content type

+++ internal/models/balance/balance.go
@@ -1,6 +1,6 @@
 package balance
 
 type Balance struct {
-	Current   float64 `json:"current"`
+	Current   float32 `json:"current"`
 	Withdrawn int     `json:"withdrawn"`
 }
